+++ .docker/php/Dockerfile
@@ -1,4 +1,5 @@
 FROM php:7.2.2-apache
-
+# enable mod_rewrite
+RUN a2enmod rewrite
 # PHP extensions
 RUN docker-php-ext-install pdo_mysql exif
