+++ .gitignore
@@ -1,5 +1,4 @@
 # Add any directories, files, or patterns you don't want to be tracked by version control
-/.htaccess
 /.docker/data
 config.php
 .DS_Store
