--- .htaccess.default
@@ -1,16 +1,0 @@
-# hide hidden files like .user.ini
-RewriteRule \/\. / [R]
-RewriteRule ^\. / [R]
-
-#BEGIN SSL and NON-WWW force
-#<IfModule mod_headers.c>
-#  RewriteEngine On
-#RewriteCond %{SERVER_PORT} 80
-#RewriteRule ^(.*)$ https://example.org/$1 [R=301,L]
-#</IfModule>
-#<IfModule mod_rewrite.c>
-#    RewriteCond %{HTTPS} !=on
-#    RewriteCond %{HTTP_HOST} ^www\.(.+)$ [NC]
-#    RewriteRule ^ https://%1%{REQUEST_URI} [R=301,L]
-#</IfModule>
-#END SSL and NON-WWW force
