+++ library/lib/errorhandling.php
@@ -1,53 +1,21 @@
 <?php
 
-function exception_error_handler($errno, $errstr, $errfile = '', $errline = '') {
-    switch ($errno) {
-	    case E_USER_ERROR:
-	    case E_USER_WARNING:
-	    case E_USER_NOTICE:
+function bootstrap_exception_handler(\Throwable $ex)
+{
+	Sentry\captureException($ex);
+	$eventId = Sentry\State\Hub::getCurrent()->getLastEventId();
 
-			$error = new Zmarty;
-			$error->assign('error',$errstr);
-			
-			if($_SESSION['user']['is_admin']) {
-				$error->assign('line',$errline);
-				$error->assign('file',$errfile);
-				$error->assign('backtrace',debug_backtrace_string());
-			}
-
-			$error->display('cms_error.tpl');
-
-			die();
-
-		case E_WARNING:
-		case E_NOTICE:
-		case E_STRICT:
-	        break;
-
-	    default:
-			echo "<h1>Something went wrong</h1>";
-	        echo '<h3>Error</h3>'.$errstr.' ('.$errno.')<br />file: '.$errfile.'<br />line: '.$errline;
-			logfile('ERROR: '.$errstr.', line '.$errline.', file '.$errfile);
-	        exit($errno);
-	        break;
-    }
+	$error = new Zmarty;
+	$error->assign('error',"Sorry, an unexpected error occured and has been reported. Please quote Sentry #$eventId.");
+	if (substr($_SERVER['HTTP_HOST'],0,strlen("localhost")) === "localhost") {
+		$error->assign('exception',$ex);
+	}
+	$error->display('cms_error.tpl');
+	die();
 }
 
-function debug_backtrace_string() {
-    $stack = '';
-    $i = 1;
-    $trace = debug_backtrace();
-    unset($trace[0]); //Remove call to this function from stack trace
-    foreach($trace as $node) {
-        $stack .= "<p>#$i ".(isset($node['file'])?$node['file']:'') ."(" .(isset($node['line'])?$node['line']:'')."): ";
-        if(isset($node['class'])) {
-            $stack .= $node['class'] . "->";
-        }
-        $stack .= $node['function'] . "()" . '</p>';
-        $i++;
-    }
-    return $stack;
-}
+set_exception_handler('bootstrap_exception_handler');
+smarty::muteExpectedErrors();
 
 function logfile($content) {
 	$content = str_replace("<br />"," ",$content);
@@ -59,6 +27,3 @@ function dump(&$var) {
 	var_export($var);
 	echo '</pre>';
 }
-
-set_error_handler("exception_error_handler");
-smarty::muteExpectedErrors();
