+++ pdf/bicyclecard.php
@@ -1,9 +1,6 @@
 <?php
 require(__DIR__.'/../library/core.php');
 
-ini_set('display_errors',1);
-error_reporting(E_ALL);
-
 require('pdf-card.php');
 
 define('PDFSCRIPT',true);
